refactor: 💡 put classes into its own directory
