+++ app/libs/__init__.py
@@ -1,7 +1,5 @@
-from .viva_application import VivaApplication
-from .my_pages import MyPages
+from .classes import Session, Viva, VivaMyPages, VivaApplication
 
 from .personal_number_helper import hashids_instace, hash_to_personal_number
-
 from .datetime_helper import milliseconds_to_date_string
 from .authenticate_helper import authenticate
