+++ app/libs/classes/__init__.py
@@ -1,0 +1,4 @@
+from .session import Session
+from .viva import Viva
+from .viva_my_pages import VivaMyPages
+from .viva_application import VivaApplication
