+++ app/routes/my_pages.py
@@ -1,7 +1,7 @@
 from flask_restful import Resource
 from zeep.exceptions import Fault
 
-from ..libs import MyPages as VivaMyPages
+from ..libs import VivaMyPages
 from ..libs import hash_to_personal_number
 
 from .. import data
